A somewhat radical redesign of the VM
A bit of development log.

I decided to abandon the idea of having the context not own the vector
of bytecode instructions. My idea of sharing instructions between many
contexts was a little misguided at this point, so that's ok.

I also decided to abandon the idea of HashMap of functions and simply
store the whole code in a single array.

With that, I had to stop storing the "current function" as a slice
(since I can't store both a Vec and a slice to it in the same data
structure).

I have src/ast.rs laying here uncomitted, but it doesn't really have
anything important right now.

Oh, and I gutted the main function. I'm doing cargo tests now.

I asked those questions in public forums:

https://stackoverflow.com/q/31067031
https://stackoverflow.com/q/31103364
https://stackoverflow.com/q/31123882
https://www.reddit.com/r/rust/comments/3bkmuh