Use debug!, Add instructions, add test cases
I grouped a couple of changes here. One is that I'm using `debug!`
instead of `println!`, so in order to show output one needs to set
`RUST_LOG=pj` in the env. But in the commited code I don't actually
initialize env_logger, so one would also need to do something like

    env_logger::init().unwrap();

Well, anyway, I added some jump conditions and binary operators, I'm
not sure if they are all useful.

Also, the semantics of jump were relaxed: now one can jump to the
point immediately after the function, which is the same as a return.

There are a number of tests, but the coverage is still weak.